+++ CHANGELOG.md
@@ -12,14 +12,24 @@ New Features
 - Added Spergel(2010) profile GSObject (#616).
 - Enable initializing a DES_PSFEx object using a pyfits HDU directly instead
   of a filename. (#626)
+- Added TopHat class implementing a circular tophat profile. (#639)
+
 
 Bug Fixes and Improvements
 --------------------------
 
+- Changed the implementation of drawing Box and Pixel profiles in real space
+  (i.e. without being convolved by anything) to actually draw the surface 
+  brightness at the center of each pixel.  This is what all other profiles do,
+  but had not been what a Box or Pixel did. (#639)
 - Fixed a bug where InterpolatedImages were not correctly rendered when
   transformed by something that includes a flip. (#645)
 - Fixed a bug if drawImage was given odd nx, ny parameters, the drawn profile
   was not correctly centered in the image. (#645)
 
+
 Updates to config options
 -------------------------
+
+- Added TopHat type. (#639)
+